+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added 
 
-- Update notebook route;
+- Added update notebook route;
+- Added notes routes;
 
 ## [0.3.0] - 2021-03-29